Refactor of the validity checking into DMSBaseMixin
WIP #735
